[FIX] Assume source_start and source_end in ms instead of frames. Closes #8.
Fixed the definition of `source_start` and `source_end` to be of type `int` instead of `float`.

Closes #8.

+++ multimodalhugs/data/datasets/pose2text.py
@@ -32,8 +32,8 @@ class Pose2TextDataset(datasets.GeneratorBasedBuilder):
     def _info(self):
         dataset_features = {
                 "source": str,
-                "source_start": Optional[float],
-                "source_end": Optional[float],
+                "source_start": Optional[int],
+                "source_end": Optional[int],
                 "source_prompt": Optional[str],
                 "generation_prompt": Optional[str],
                 "output_text": Optional[str],
